app/lambdas/requirements.txt
praw
python-dotenv
boto3